refactor(validate): inline body-flag loop, drop BodyFlags helper
Removes internal/validate.BodyFlags entirely. The two callers
(decision add, learning add) now loop their body flags themselves
in PreRunE and call validate.RejectPlaceholder per (flag, value)
pair. The wiring is fully visible at the noun-level constructor
and the validate package becomes uniformly string-consuming
(matches Symlinks, RejectPlaceholder); the cobra.Command-taking
outlier is gone.

The RejectPlaceholder helper moves to its own file
(rejectplaceholder.go / rejectplaceholder_test.go) per the
single-helper-per-file convention used elsewhere in the project
(internal/sanitize/truncate.go). The BodyFlags fixture and tests
that needed a *cobra.Command go away with it; only the four
RejectPlaceholder unit tests remain.

Bundled in this commit because they share the same branch arc
and the repo is mid-rebuild after a Go toolchain bump:

* go.mod bumped 1.26.1 -> 1.26.3 to match the local toolchain
  (the 1.26.1 cached toolchain in ~/go/pkg/mod was corrupt).
* SKILL.md (ctx-spec copilot integration) gains the --brief
  contract section that was already merged into the canonical
  /ctx-spec skill in 55acbd81 but missed the embedded asset.
* Handover note from the prior session is preserved under
  .context/handovers/ following the existing precedent.

Spec: specs/skill-surface-polish.md
