+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/ActiveMemory/ctx
 
-go 1.26.1
+go 1.26.3
 
 require (
 	github.com/hashicorp/raft v1.7.3
