+++ internal/cli/learning/cmd/add/cmd.go
@@ -21,19 +21,32 @@ import (
 // Adds a new learning entry to LEARNINGS.md with the
 // required provenance, context, lesson, and application
 // flags. Implementation lives in the shared add core; this
-// noun-level constructor installs a PreRunE that calls
-// [validate.BodyFlags] to reject empty or placeholder values
-// on the three body flags.
+// noun-level constructor installs a PreRunE that reads each
+// body flag and calls [validate.RejectPlaceholder] to reject
+// empty or placeholder values.
 //
 // Returns:
 //   - *cobra.Command: Configured learning add subcommand
 func Cmd() *cobra.Command {
 	c := build.Cmd(entry.Learning, cmd.DescKeyLearningAdd, cmd.UseLearningAdd)
 	c.PreRunE = func(cobraCmd *cobra.Command, _ []string) error {
-		return validate.BodyFlags(
-			cobraCmd,
-			cFlag.Context, cFlag.Lesson, cFlag.Application,
-		)
+		flags := cobraCmd.Flags()
+		names := []string{
+			cFlag.Context,
+			cFlag.Lesson,
+			cFlag.Application,
+		}
+		for _, name := range names {
+			value, getErr := flags.GetString(name)
+			if getErr != nil {
+				return getErr
+			}
+			rejectErr := validate.RejectPlaceholder(name, value)
+			if rejectErr != nil {
+				return rejectErr
+			}
+		}
+		return nil
 	}
 	return c
 }
